api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
chromadb>=0.5.0
litellm>=1.50.0
sentence-transformers>=3.0.0
python-multipart>=0.0.12
langchain-text-splitters>=0.3.0
pypdf>=5.0.0
python-dotenv>=1.0.0
google-auth>=2.0.0
pyjwt>=2.0.0
